refactor:  move app.js to engine.js & application.js
