+++ app/extend/application.js
@@ -1,0 +1,36 @@
+'use strict';
+
+const VIEW_ENGINE = Symbol('app#ViewEngine');
+const VIEW_HELPER = Symbol('app#ViewHelper');
+const View = require('../../lib/view');
+const helper = require('../../lib/helper');
+const engine = require('../../lib/engine');
+
+module.exports = {
+  // mount `View` class to app
+  // egg will create an instance to `ctx.view` at every request
+  // you can use `this.render` at controller
+  get [Symbol.for('egg#view')]() {
+    return View;
+  },
+
+  /**
+   * nunjucks environment
+   * @member {Object} Application#viewEngine
+   * @see https://mozilla.github.io/nunjucks/api.html#environment
+   */
+  get viewEngine() {
+    if (!this[VIEW_ENGINE]) {
+      this[VIEW_ENGINE] = engine(this);
+    }
+    return this[VIEW_ENGINE];
+  },
+
+  // view helper
+  get ViewHelper() {
+    if (!this[VIEW_HELPER]) {
+      this[VIEW_HELPER] = helper(this);
+    }
+    return this[VIEW_HELPER];
+  },
+};
