+++ config/config.default.js
@@ -8,7 +8,7 @@ module.exports = function(appInfo) {
   /**
    * View options
    * @member Config#view
-   * @property {String} loadpath - full path of template, default to `app/views`, support multi with comma
+   * @property {String} loadpath - full path of templates, defaults to `app/views`, multiple paths are supported when paths separated by commas.
    * @property {Boolean} cache - whether cache template, default to true but false at local env.
    */
   exports.view = {
