+++ lib/engine.js
@@ -1,10 +1,14 @@
 'use strict';
 
+const path = require('path');
 const nunjucks = require('nunjucks');
 
 module.exports = app => {
+  const coreLogger = app.loggers.coreLogger;
   const config = app.config.view;
   const viewPaths = config.loadpath.split(',');
+
+  coreLogger.info('[egg:plugin:view] loading templates from %j', viewPaths);
 
   const options = Object.assign({
     noCache: config.cache !== true,
@@ -14,6 +18,9 @@ module.exports = app => {
   const engine = new nunjucks.Environment(getFileLoader(options), options);
 
   engine.nunjucks = nunjucks;
+
+  // filter loader
+  loadFilter();
 
   /**
    * clean template cache
@@ -64,6 +71,16 @@ module.exports = app => {
       return result;
     };
     return fileLoader;
+  }
+
+  // load `app/extend/filters.js` from app/framework/plugin into nunjucks
+  function loadFilter() {
+    for (const dir of app.loader.loadDirs()) {
+      const filters = app.loader.loadFile(path.join(dir, 'app/extend/filter.js')) || {};
+      for (const name of Object.keys(filters)) {
+        engine.addFilter(name, filters[name]);
+      }
+    }
   }
 
   return engine;
